Filtering while searching
This refers to goyalsid/phageParser#47

I have added a new script, in order to leave the previous one as it is. In order to make debugging easier, it takes two parameters, email and accession number this way:

acc_and_filter.py youre@ma.il data/bac_accession_list_tiny.txt

This second file includes one accession number with Cas3 and others. And for the time being it just prints the product in case it finds Cas3 and others. I don't know Python, have worked only occassionally with `Bio` (and usually in Perl and FASTA) and this was just a proof of concept.

As you see, I use Bio.Seq. It's probably a more sensible choice to do it just the way you did it before: producing gb files and then dealing with them; but you can just as well work on the humongous GB file using Bio.SeqIO (in Python) or Bio::SeqIO (in Perl). No need at all to use regular expressions. My point is that by looking at how you run over the records it should be easy to rewrite either cas_in_gb (using dependencies, no other way) or do it in Python reading from file instead of using `efetch`

At any rate, check it now to see if it's close to what you're looking for and we can take it from there, making output resemble what we had before with the Perl script or rewriting the Perl script (in Perl or Python) so that it acts on the humongous example file.

Don't pull request now, but you can pull it form my copy of the repository.